# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Modern Agent HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/companies:
    get:
      operationId: listCompanies
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCompaniesResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List all registered companies
      tags:
      - companies
    post:
      operationId: createCompany
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Register a new company
      tags:
      - companies
...
  /api/v1/notifications:
    get:
      operationId: listNotifications
      parameters:
      - description: Notification status filter. V1 supports only unread.
        in: query
        name: status
        schema:
          description: Notification status filter. V1 supports only unread.
          enum:
          - unread
          type: string
      - description: Maximum notifications to return. Defaults to 50; capped at 100.
        in: query
        name: limit
        schema:
          description: Maximum notifications to return. Defaults to 50; capped at
            100.
          maximum: 100
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListNotificationsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List unread notifications
      tags:
      - notifications
  /api/v1/notifications/{id}:
    patch:
      operationId: markNotificationRead
      parameters:
      - description: Notification identifier.
        in: path
        name: id
        required: true
        schema:
          description: Notification identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkNotificationReadRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationEnvelope'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Mark a notification read
      tags:
      - notifications
  /api/v1/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAllNotificationsReadResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Mark all unread notifications read
      tags:
      - notifications
  /api/v1/notifications/stream:
    get:
      operationId: streamNotifications
      parameters:
      - description: Optional project id filter for live notifications.
        in: query
        name: projectId
        schema:
          description: Optional project id filter for live notifications.
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Stream created notifications
      tags:
      - notifications
...
  /api/v1/projects/{id}/company:
    put:
      operationId: assignProjectCompany
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignProjectCompanyInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignProjectCompanyResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Assign (or, with an empty companyId, unassign) a project's company
      tags:
      - companies
...
  /api/v1/projects/{id}/messages:
    get:
      operationId: listProjectMessages
      parameters:
      - description: Project identifier (registry key).
        in: path
        name: id
        required: true
        schema:
          description: Project identifier (registry key).
          type: string
      - description: Maximum messages to return. Defaults to 100; capped at 500.
        in: query
        name: limit
        schema:
          description: Maximum messages to return. Defaults to 100; capped at 500.
          maximum: 500
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectMessagesResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List durable agent-to-agent send messages targeting a project's sessions
      tags:
      - sessions
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    AssignProjectCompanyInput:
      properties:
        companyId:
          type: string
      required:
      - companyId
      type: object
    AssignProjectCompanyResponse:
      properties:
        companyId:
          type: string
        projectId:
          type: string
      required:
      - projectId
      type: object
...
    Company:
      properties:
        createdAt:
          format: date-time
          type: string
        id:
          type: string
        name:
          type: string
      required:
      - id
      - name
      - createdAt
      type: object
    CompanyResponse:
      properties:
        company:
          $ref: '#/components/schemas/Company'
      required:
      - company
      type: object
...
    CreateCompanyInput:
      properties:
        name:
          type: string
      required:
      - name
      type: object
...
    ListCompaniesResponse:
      properties:
        companies:
          items:
            $ref: '#/components/schemas/Company'
          type: array
      required:
      - companies
      type: object
    ListNotificationsResponse:
      properties:
        notifications:
          items:
            $ref: '#/components/schemas/NotificationResponse'
          type: array
      required:
      - notifications
      type: object
    ListProjectMessagesResponse:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/SessionMessage'
          type: array
      required:
      - messages
      type: object
...
    MarkAllNotificationsReadResponse:
      properties:
        notifications:
          items:
            $ref: '#/components/schemas/NotificationResponse'
          type: array
      required:
      - notifications
      type: object
    MarkNotificationReadRequest:
      properties:
        status:
          description: V1 supports only marking an unread notification read.
          enum:
          - read
          type: string
      required:
      - status
      type: object
...
    NotificationEnvelope:
      properties:
        notification:
          $ref: '#/components/schemas/NotificationResponse'
      required:
      - notification
      type: object
    NotificationResponse:
      properties:
        body:
          type: string
        createdAt:
          format: date-time
          type: string
        id:
          type: string
        prUrl:
          type: string
        projectId:
          type: string
        sessionId:
          type: string
        status:
          enum:
          - unread
          - read
          type: string
        target:
          $ref: '#/components/schemas/NotificationTarget'
        title:
          type: string
        type:
          enum:
          - needs_input
          - ready_to_merge
          - pr_merged
          - pr_closed_unmerged
          - auto_terminated
          type: string
      required:
      - id
      - sessionId
      - projectId
      - prUrl
      - type
      - title
      - body
      - status
      - createdAt
      - target
      type: object
    NotificationTarget:
      properties:
        kind:
          enum:
          - session
          - pr
          type: string
        prUrl:
          type: string
        sessionId:
          type: string
      required:
      - kind
      - sessionId
      type: object
...
    SessionMessage:
      properties:
        content:
          type: string
        createdAt:
          format: date-time
          type: string
        id:
          type: string
        senderSessionId:
          type: string
        targetSessionId:
          type: string
      required:
      - id
      - targetSessionId
      - content
      - createdAt
      type: object
...